(lmwang)delete run_node from results_manager only when it exists
git-svn-id: https://svn.urbansim.org/src/trunk@16889 1513fa5d-83ce-422e-ae47-6967bb0fb594

+++ opus_gui/results_manager/results_manager_functions.py
@@ -59,7 +59,8 @@ def get_available_run_nodes(project):
 
 def delete_simulation_run(project, run_name):
     run_node = project.find('results_manager/simulation_runs', name=run_name)
-    get_manager_instance('results_manager').delete_run(run_node)
+    if run_node:  # delete the run_node only when it exists
+        get_manager_instance('results_manager').delete_run(run_node)
 
 def add_simulation_run(project, cache_directory, scenario_name, run_name,
                        start_year, end_year, run_id):
